Fix tests

+++ runtime/src/cached_executors.rs
@@ -437,7 +437,7 @@ pub(crate) mod tests {
     impl Executor for TestExecutor {
         fn execute(
             &self,
-            _first_instruction_account: usize,
+            _first_instruction_account: u16,
             _invoke_context: &mut InvokeContext,
         ) -> std::result::Result<(), InstructionError> {
             Ok(())
